requirements.txt
mysql-connector == 2.2.9

python-dotenv~=1.0.0